refactor: Update BeneficiaryController and BeneficiaryService
This commit refactors the `BeneficiaryController` and `BeneficiaryService` classes. In the `BeneficiaryController`, the `findAll` method no longer accepts any parameters. In the `BeneficiaryService`, the `findAll` method now retrieves data from the database using the `rsprisma` client and sends it to the `client` using the `send` method. This change improves the separation of concerns and simplifies the code.